+++ src/rs.cpp
@@ -417,7 +417,7 @@ double rs_get_detached_frame_metadata(const rs_frame_ref * frame, rs_frame_metad
 }
 HANDLE_EXCEPTIONS_AND_RETURN(0, frame)
 
-double rs_supports_frame_metadata(const rs_frame_ref * frame, rs_frame_metadata frame_metadata, rs_error ** error) try
+int rs_supports_frame_metadata(const rs_frame_ref * frame, rs_frame_metadata frame_metadata, rs_error ** error) try
 {
     VALIDATE_NOT_NULL(frame);
     return frame->supports_frame_metadata(frame_metadata);
